Have buildApp and runEndToEnd use git libraries
There is an open issue where GIT_COMMIT / GIT_BRANCH etc are not exposed
to the github-branch-source and therefore not available from Jenkins
multibranchPipelines. We have added libraries in previous commit to get/
set this information. This commit changes the buildApp and runEndToEnd
libraries to use the gitCommit and gitBranch library to achieve the same
result.