PP-3527 Fix smoke test calls
There is now a base runSmokeTest library fn that calls
the `run-smoke-test` Jenkins job. That job uses the new
version of the smoke script (run-smokev2) if you pass it
`SMOKE_TAGS`. So all this works together, although a tad
complicated to trace through. Once we've transitioned over
we can delete all the old stuff and it will be clearer.