PP-3776 Temporarily disable direct debit smoke tests
DD connector is broken, and DD team want to take some time to
diafnose the issue. While they do that we are disabling this smoke
test, so as not to block the pipeline. This will mean that
DD smoke tests won't be run post deploy to test, so some apps may
get to staging in a 'broken' state (though this is unlikely). If
they do, staging smoke tests will catch them.