bau: Specify a flag to publish verification results to the pact broker
We have (or will be upgrading) java projects to use version 3.5.18 of the pact
jvm provider library, which disables publishing of verification result to the
pact broker by default. See https://github.com/DiUS/pact-jvm/issues/695. We
want to do this so results don't get accidentally published from development
machines.So we have to set the pact.verifier.publishResults to true here.

@oswaldquek