bau: Publish verification when running provider test in a consumer build
This resolves an issue where a new interaction is added in the consumer pact,
but does not require a change in the provider. When this happens, when the
provider contract test is run as part of the consumer build, the verification
result should be published. Otherwise, the consumer build will fail at the
deploy stage as no provider has verified the new interaction that was added in
the first place.

@oswaldquek