Tag repositories inline in their build jobs
Rather than call out to `run-tag-and-capture-notes-commit-based` as a separate
job, which clones a repo of scripts and then clones the repo of the application
we want to tag, we can just tag and push the repository we are within, inline
within the build job.

Convert our tagging script to groovy and inline it.